Split out SecretMenu from MenuScreen, relax warnings inside SecretMenu
